% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Params.R
\name{ExperimentParam}
\alias{ExperimentParam}
\title{Experiment setup}
\usage{
ExperimentParam(
  wd = ".",
  d.res = file.path(wd, "Result"),
  d.tmp = file.path(d.res, "tmp"),
  element.trace = "12C",
  element.label = "13C",
  equipment = c("QTOF", "Obiwarp"),
  resolution = NULL,
  ppm = 25,
  res.define = 400,
  nSlaves = 4
)
}
\arguments{
\item{wd}{Working directory for experiment to be processed}

\item{d.res}{Folder for storing results}

\item{d.tmp}{Folder for storing files when processing. By default, it is the
same with \code{wd}}

\item{element.trace}{Monoisotope elements, e.g., 12C or 14N}

\item{element.label}{Isotope labelled elements, e.g., 13C or 15N}

\item{equipment}{Equipment type: QTOF/Obiwarp}

\item{resolution}{Resolution of equipment, for QTOF, it is 3000 by default}

\item{ppm}{Mass to charge value (m/z) tolerance for MS1 ions}

\item{res.define}{Where the PPP resolution starts from. For QTOF equipments
in zhulab, it is set to 400}

\item{nSlaves}{Number of threads to be used}
}
\value{
an \code{ExperimentParam} object
}
\description{
Experiment setup
}
