% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractIsotopicPeaks.R
\name{ExtractIsotopeEIC}
\alias{ExtractIsotopeEIC}
\title{Extract EIC of isomers}
\usage{
ExtractIsotopeEIC(xr, mzrange, mz.isomer = NULL, rtrange = NULL)
}
\arguments{
\item{xr}{`xcmsSet` xcmsSet dataset from xcms}

\item{mzrange}{`list` mz ranges for extraction}

\item{mz.isomer}{`list` mz values for isomers}

\item{rtrange}{`list` rt ranges for extraction}
}
\value{
list EICs
}
\description{
Extract EIC of isomers according to their mz range cross all scans and
return the mzs and intensities of the EIC as well as the scantime for the
file
}
