% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Params.R
\name{ExtractParam}
\alias{ExtractParam}
\title{Isotopoluage extraction parameter setup}
\usage{
ExtractParam(
  d.extract = "extract",
  range.extract = c("all", "roi", "rtrange"),
  method.align = c("apex", "eic"),
  method.best = c("maxint", "maxcor"),
  method.quant = c("rawinroi", "raw", "baseline"),
  nscan.quant = 3,
  minfrac = 0.5,
  adj.unlabel = NULL,
  adj.label = NULL,
  adj.contaminate = FALSE,
  thr.contaminate = 0.02
)
}
\arguments{
\item{d.extract}{directory of samples to be extracted}

\item{range.extract}{EIC extraction range
\itemize{
    \item[] 'all' - extract the entire eic
    \item[] 'roi' - extract the rois related to rt range of target peaks
    \item []'rtrange' - extract the eic at the range of rt range of target peaks
    }}

\item{method.align}{method for creating representative peak profile
\itemize{
    \item[] 'apex' align with the apex of the most similar peaks
    \item[] 'eic' align with the correlation of the eics from unlabelled and
          labelled samples
}}

\item{method.best}{method for finding best correlated peak groups for extracted
   isotopoluage to determine the rt range of extracted peaks/eics
\itemize{
    \item[] 'maxint' select the highest peak as rt range reference
    \item[] 'maxcor select the most similar peak to representive peak as rt range reference
}
@param method.quant method for quantifying extracted isotopic peaks.
\itemize{
    \item[] 'rawinroi' sum with raw intensities of significant scans (in ROI)
    \item[] 'baseline' baseline substraxted rawinroi method
    \item[] 'raw' sum with raw intensities anyway
}}

\item{nscan.quant}{number of scans to sum up for quantifying}

\item{minfrac}{minimal fraction threshold to determine the isotopic peak
existed or not in the sample groups}

\item{adj.unlabel}{character vector, group names of unlabelled samples for
adjusting the isotopoluage of labelled samples}

\item{adj.label}{character vector, group names of labelled samples to be
adjusted using the unlabelled samples (pairwised with the sample groups in
\code{adj.unlabel})}

\item{adj.contaminate}{logical: if adjust the contaminate using unlabelled
samples, if \code{TRUE}, \code{adj.unlabel} and \code{adj.label} must be set}
}
\value{
an \code{ExtractParam} object
}
\description{
Isotopoluage extraction parameter setup
}
