% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Params.R
\name{IsotopologueParam}
\alias{IsotopologueParam}
\title{Isotopoluage target generation parameter setup}
\usage{
IsotopologueParam(
 
    file.id = "unlabelled/MetProcess-Result/MRN_annotation_result/MRN.annotation.result.csv",
  file.xset = "unlabelled/xset-centWave-final.Rda",
  method = c("max", "sum", "median"),
  value = c("into", "intb", "maxo"),
  rt.extend = 30,
  is.plot.pattern = FALSE
)
}
\arguments{
\item{file.id}{MetDNA annotation result file}

\item{file.xset}{xcms peak detection file}

\item{method}{method for creating representative peak profile, (max, sum, median)}

\item{value}{value field for creating representative peak profile (into, intb, maxo)}

\item{rt.extend}{extention thresholds for RT}

\item{is.plot.pattern}{if plot isotopoluage patterns}
}
\value{
an \code{extractTargetsParam} object
}
\description{
Isotopoluage target generation parameter setup
}
