% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{LoadData}
\alias{LoadData}
\title{Load data to envirement by assigning it to a new object}
\usage{
LoadData(file, keep.name = FALSE, env)
}
\description{
Load data to envirement by assigning it to a new object
}
