% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Params.R
\name{PeakdetectionParam}
\alias{PeakdetectionParam}
\title{Peak detection parameter setup}
\usage{
PeakdetectionParam(
  peakwidth = c(5, 30),
  snthr = 3,
  prefilter = c(3, 100),
  fitgauss = FALSE,
  method.peakdetection = c("centWave", "localMax"),
  method.roi = c("continuous", "aroundMax"),
  method.baseline = c("xcms", "centWaveP"),
  method.smooth = c("Gaussian", "LOESS", "SG")
)
}
\arguments{
\item{peakwidth}{`numeric(2)` with the lower and upper bound of the
expected peak width.}

\item{snthr}{`numeric(1)` defining the signal to noise ratio cutoff.
Peaks with a signal to noise ratio < `snthr` are omitted.}

\item{prefilter}{`numeric(2)` (`c(k, I)`): only regions of interest with at
least `k` centroids with signal `>= I` are returned in the first
step.#'}

\item{fitgauss}{`logical(1)` whether or not a Gaussian should be fitted
to each peak.}

\item{method.peakdetection}{`character(1)` method for peak detection}

\item{method.roi}{method for finding rois in extracted EICs.
\itemize{
    \item[] 'continuous' - traditional centwave roi finding method
    \item[] 'aroundMax' - finding local maximums and determing a scan range
     based on the peakwidth to define the roi range
}}

\item{method.baseline}{method for determing baselines, only xcms supported currently}

\item{method.smooth}{`character(1)` method for smoothing the EICs (Gaussian, LOESS and SG)}
}
\value{
an \code{PeakdetectionParam} object
}
\description{
Peak detection parameter setup
}
